refactor Recoil atoms to reduce file numbers, update imports, update READMEs
