+++ src/components/pages/GameContainer/Game/Postgame/Postgame.tsx
@@ -7,7 +7,7 @@ import {
   loadingState,
   lobbyState,
   playerGuessState,
-} from '../../../../../state';
+} from '../../../../../state/gameState';
 import {
   DefinitionResultItem,
   GuessItem,
