--- src/state/availableReactionsState.ts
@@ -1,7 +1,0 @@
-import { atom } from 'recoil';
-import { ReactionItem } from '../types/commonTypes';
-
-export const availableReactionsState = atom<ReactionItem[]>({
-  key: 'availableReactionsState',
-  default: [],
-});
