--- src/state/definitionReactionsState.ts
@@ -1,7 +1,0 @@
-import { atom } from 'recoil';
-import { ReactionsDictionary } from '../types/gameTypes';
-
-export const definitionReactionsState = atom<ReactionsDictionary>({
-  key: 'definitionReactionsState',
-  default: {},
-});
