--- src/state/guessesState.ts
@@ -1,7 +1,0 @@
-import { atom } from 'recoil';
-import { GuessItem } from '../types/gameTypes';
-
-export const guessesState = atom<GuessItem[]>({
-  key: 'guessesState',
-  default: [],
-});
