--- src/state/handleSendReactionFn.ts
@@ -1,10 +1,0 @@
-import { atom } from 'recoil';
-
-export const handleSendReactionFn = atom<HandleSendReactionFn>({
-  key: 'handleSendReactionFn',
-  default: () => null,
-});
-
-type HandleSendReactionFn =
-  | (() => null)
-  | ((definitionId: number, reactionId: number) => void);
