--- src/state/lobbyCodeState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const lobbyCodeState = atom<string>({
-  key: 'lobbyCode',
-  default: '',
-});
