--- src/state/playerIdState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const playerIdState = atom<string>({
-  key: 'playerId',
-  default: '',
-});
