--- src/state/revealResultsState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const revealResultsState = atom<boolean>({
-  key: 'revealResultsState',
-  default: false,
-});
