--- src/state/showNewHostModalState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const showNewHostModalState = atom<boolean>({
-  key: 'showNewHostModalState',
-  default: false,
-});
