--- src/state/tokenState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const tokenState = atom<string>({
-  key: 'token',
-  default: '',
-});
