+++ src/types/gameTypes.ts
@@ -1,5 +1,11 @@
 import { DoubleNumberDict } from './commonTypes';
 
+/* Types */
+
+export type LoadingState = 'ok' | 'loading' | 'failed';
+
+/* Interfaces */
+
 export interface PlayerItem {
   id: string;
   username: string;
@@ -127,4 +133,9 @@ export interface GetReactionsItem {
   value: number;
   definition_id: number;
   reaction_id: number;
+}
+
+export interface Timer {
+  startTime: number;
+  currentTime: number;
 }
